Add the beat clock

+++ src/sfizz.hpp
@@ -344,7 +344,7 @@ public:
      * @param bar     The current bar.
      * @param barBeat The fractional position of the current beat within the bar.
      */
-    void timePosition(int delay, int bar, float barBeat);
+    void timePosition(int delay, int bar, double barBeat);
 
     /**
      * @brief Send the playback state.
